requirements.txt
django
gunicorn
djangorestframework
markdown       
django-filter
drf-writable-nested
django-url-filter
psycopg2